+++ astrodata/doc/ad_CheatSheet/conf.py
@@ -63,7 +63,7 @@ version = '3.0'
 # The full version, including alpha/beta/rc tags.
 #release = '3.0.x'
 #rtdurl = 'release-'+release
-release = '3.0.2'
+release = '3.0.3'
 rtdurl = 'v'+release
 
 
@@ -265,10 +265,10 @@ texinfo_documents = [
 # Example configuration for intersphinx: refer to the Python standard library.
 # Example configuration for intersphinx: refer to the Python standard library.
 intersphinx_mapping = {
-    'astrodata': ('https://astrodata-user-manual.readthedocs.io/en/v3.0.2/', None),
+    'astrodata': ('https://astrodata-user-manual.readthedocs.io/en/v3.0.3/', None),
     'astropy': ('https://docs.astropy.org/en/stable/', None),
-    'gemini_instruments': ('https://dragons-recipe-system-programmers-manual.readthedocs.io/en/v3.0.2/', None),
-    'geminidr': ('https://dragons-recipe-system-programmers-manual.readthedocs.io/en/v3.0.2/', None),
+    'gemini_instruments': ('https://dragons-recipe-system-programmers-manual.readthedocs.io/en/v3.0.3/', None),
+    'geminidr': ('https://dragons-recipe-system-programmers-manual.readthedocs.io/en/v3.0.3/', None),
     'matplotlib': ('https://matplotlib.org/', None),
     'numpy': ('https://numpy.org/doc/stable/', None),
     'python': ('https://docs.python.org/3/', None),