+++ geminidr/doc/tutorials/GMOSImg-DRTutorial/conf.py
@@ -42,7 +42,7 @@ version = '3.0'
 #release = astrodata.version()
 #release = '3.0.x'
 #rtdurl = 'release-'+release
-release = '3.0.2'
+release = '3.0.3'
 rtdurl = 'v'+release
 
 
@@ -214,10 +214,10 @@ epub_exclude_files = ['search.html']
 
 # Example configuration for intersphinx: refer to the Python standard library.
 intersphinx_mapping = {
-    'astrodata': ('https://astrodata-user-manual.readthedocs.io/en/v3.0.2/', None),
+    'astrodata': ('https://astrodata-user-manual.readthedocs.io/en/v3.0.3/', None),
     'astropy': ('http://docs.astropy.org/en/stable/', None),
-    'gemini_instruments': ('https://dragons-recipe-system-programmers-manual.readthedocs.io/en/v3.0.2/', None),
-    'geminidr': ('https://dragons-recipe-system-programmers-manual.readthedocs.io/en/v3.0.2/', None),
+    'gemini_instruments': ('https://dragons-recipe-system-programmers-manual.readthedocs.io/en/v3.0.3/', None),
+    'geminidr': ('https://dragons-recipe-system-programmers-manual.readthedocs.io/en/v3.0.3/', None),
     'matplotlib': ('https://matplotlib.org/', None),
     'numpy': ('https://docs.scipy.org/doc/numpy/', None),
     'python': ('https://docs.python.org/3', None),