+++ geminidr/doc/tutorials/GMOSLS-DRTutorial/conf.py
@@ -25,7 +25,7 @@ author = 'Kathleen Labrie'
 version = '3.0'
 #release = '3.0.x'
 #rtdurl = 'release-'+release
-release = '3.0.2'
+release = '3.0.3'
 rtdurl = 'v'+release
 
 today = 'July 2022'